chore: bulk commit all multi-agent work — routing, synthesis, scoring, UI, CI, cleanup
Major changes across all agents:
- New routing components (16 new, 14 removed/replaced) with improved token gating
- Synthesis template refactor: split into _templates_core/exotic/research/routing
- Intelligence layer: GNN predictor, interaction analysis, temporal Bayesian
- Research tools: ppl_sweep, backfill scripts, audit tools, monitor agent
- UI refactor: CanvasArea, ModalLayer, RightPanel, TopBar/ParamInput updates
- CI: governance-ci, pipeline-ci, weekly-audit GitHub workflows
- Conductor orchestration module
- Archive old artifacts and completed task plans
- Clean up MagicMock/SQLite junk files, updated .gitignore
- Fix ruff lint errors (unused vars, duplicate keys, unused imports)

+++ research/mathspaces/clifford.py
@@ -342,7 +342,7 @@ def execute_clifford_attention(module: nn.Module, x: torch.Tensor) -> torch.Tens
         n = D_padded * D_padded
         if W.numel() >= n:
             Wq = W[:n].reshape(D_padded, D_padded)
-            q = F.linear(x_padded, Wq)
+            q = F.linear(x_padded, Wq.to(x_padded.dtype))
             k = x_padded
         else:
             q = k = x_padded
